fix: Change data type of actuator constants from double to float as precision is easily sufficient

+++ include/myactuator_rmd/actuator_constants.hpp
@@ -22,10 +22,10 @@ namespace myactuator_rmd {
   */
   class X10S2V3 {
     public:
-      static constexpr double reducer_ratio {35/1};
-      static constexpr double speed_constant {30}; // in rpm/V
-      static constexpr double torque_constant {0.32}; // in Nm/A
-      static constexpr double rotor_inertia {5675}; // in gcm2
+      static constexpr float reducer_ratio {35/1};
+      static constexpr float speed_constant {30}; // in rpm/V
+      static constexpr float torque_constant {0.32}; // in Nm/A
+      static constexpr float rotor_inertia {5675}; // in gcm2
   };
 
   /**\class X4V3
@@ -35,10 +35,10 @@ namespace myactuator_rmd {
   */
   class X4V3 {
     public:
-      static constexpr double reducer_ratio {6/1};
-      static constexpr double speed_constant {200}; // in rpm/V
-      static constexpr double torque_constant {0.35}; // in Nm/A
-      static constexpr double rotor_inertia {200}; // in gcm2
+      static constexpr float reducer_ratio {6/1};
+      static constexpr float speed_constant {200}; // in rpm/V
+      static constexpr float torque_constant {0.35}; // in Nm/A
+      static constexpr float rotor_inertia {200}; // in gcm2
   };
 
   /**\class X6V3
@@ -48,10 +48,10 @@ namespace myactuator_rmd {
   */
   class X6V3 {
     public:
-      static constexpr double reducer_ratio {8/1};
-      static constexpr double speed_constant {62}; // in rpm/V
-      static constexpr double torque_constant {0.17}; // in Nm/A
-      static constexpr double rotor_inertia {850}; // in gcm2
+      static constexpr float reducer_ratio {8/1};
+      static constexpr float speed_constant {62}; // in rpm/V
+      static constexpr float torque_constant {0.17}; // in Nm/A
+      static constexpr float rotor_inertia {850}; // in gcm2
   };
 
   /**\class X8ProV2
@@ -61,10 +61,10 @@ namespace myactuator_rmd {
   */
   class X8ProV2 {
     public:
-      static constexpr double reducer_ratio {9/1};
-      static constexpr double speed_constant {30}; // in rpm/V
-      static constexpr double torque_constant {2.6}; // in Nm/A
-      static constexpr double rotor_inertia {3400}; // in gcm2
+      static constexpr float reducer_ratio {9/1};
+      static constexpr float speed_constant {30}; // in rpm/V
+      static constexpr float torque_constant {2.6}; // in Nm/A
+      static constexpr float rotor_inertia {3400}; // in gcm2
       static constexpr std::int16_t number_of_pole_pairs {20};
   };
 
@@ -75,10 +75,10 @@ namespace myactuator_rmd {
   */
   class X8S2V3 {
     public:
-      static constexpr double reducer_ratio {36/1};
-      static constexpr double speed_constant {33}; // in rpm/V
-      static constexpr double torque_constant {0.30}; // in Nm/A
-      static constexpr double rotor_inertia {2670}; // in gcm2
+      static constexpr float reducer_ratio {36/1};
+      static constexpr float speed_constant {33}; // in rpm/V
+      static constexpr float torque_constant {0.30}; // in Nm/A
+      static constexpr float rotor_inertia {2670}; // in gcm2
   };
 
   /**\class X6V2
@@ -88,10 +88,10 @@ namespace myactuator_rmd {
   */
   class X6V2 {
     public:
-      static constexpr double reducer_ratio {6/1};
-      static constexpr double speed_constant {60}; // in rpm/V
-      static constexpr double torque_constant {0.88}; // in Nm/A
-      static constexpr double rotor_inertia {800}; // in gcm2
+      static constexpr float reducer_ratio {6/1};
+      static constexpr float speed_constant {60}; // in rpm/V
+      static constexpr float torque_constant {0.88}; // in Nm/A
+      static constexpr float rotor_inertia {800}; // in gcm2
       static constexpr std::int16_t number_of_pole_pairs {14};
   };
 
@@ -102,10 +102,10 @@ namespace myactuator_rmd {
   */
   class X8V2 {
     public:
-      static constexpr double reducer_ratio {9/1};
-      static constexpr double speed_constant {40}; // in rpm/V
-      static constexpr double torque_constant {2.09}; // in Nm/A
-      static constexpr double rotor_inertia {2600}; // in gcm2
+      static constexpr float reducer_ratio {9/1};
+      static constexpr float speed_constant {40}; // in rpm/V
+      static constexpr float torque_constant {2.09}; // in Nm/A
+      static constexpr float rotor_inertia {2600}; // in gcm2
       static constexpr std::int16_t number_of_pole_pairs {21};
   };
 
@@ -116,10 +116,10 @@ namespace myactuator_rmd {
   */
   class X10V3 {
     public:
-      static constexpr double reducer_ratio {7/1};
-      static constexpr double speed_constant {30}; // in rpm/V
-      static constexpr double torque_constant {0.32}; // in Nm/A
-      static constexpr double rotor_inertia {5675}; // in gcm2
+      static constexpr float reducer_ratio {7/1};
+      static constexpr float speed_constant {30}; // in rpm/V
+      static constexpr float torque_constant {0.32}; // in Nm/A
+      static constexpr float rotor_inertia {5675}; // in gcm2
   };
 
   /**\class X4V2
@@ -129,10 +129,10 @@ namespace myactuator_rmd {
   */
   class X4V2 {
     public:
-      static constexpr double reducer_ratio {10/1};
-      static constexpr double speed_constant {12.5}; // in rpm/V
-      static constexpr double torque_constant {0.625}; // in Nm/A
-      static constexpr double rotor_inertia {105}; // in gcm2
+      static constexpr float reducer_ratio {10/1};
+      static constexpr float speed_constant {12.5}; // in rpm/V
+      static constexpr float torque_constant {0.625}; // in Nm/A
+      static constexpr float rotor_inertia {105}; // in gcm2
       static constexpr std::int16_t number_of_pole_pairs {14};
   };
 
@@ -143,10 +143,10 @@ namespace myactuator_rmd {
   */
   class X6S2V2 {
     public:
-      static constexpr double reducer_ratio {36/1};
-      static constexpr double speed_constant {135}; // in rpm/V
-      static constexpr double torque_constant {6.92}; // in Nm/A
-      static constexpr double rotor_inertia {800}; // in gcm2
+      static constexpr float reducer_ratio {36/1};
+      static constexpr float speed_constant {135}; // in rpm/V
+      static constexpr float torque_constant {6.92}; // in Nm/A
+      static constexpr float rotor_inertia {800}; // in gcm2
       static constexpr std::int16_t number_of_pole_pairs {14};
   };
 
@@ -157,10 +157,10 @@ namespace myactuator_rmd {
   */
   class X8HV3 {
     public:
-      static constexpr double reducer_ratio {6.2/1};
-      static constexpr double speed_constant {33}; // in rpm/V
-      static constexpr double torque_constant {0.30}; // in Nm/A
-      static constexpr double rotor_inertia {2670}; // in gcm2
+      static constexpr float reducer_ratio {6.2/1};
+      static constexpr float speed_constant {33}; // in rpm/V
+      static constexpr float torque_constant {0.30}; // in Nm/A
+      static constexpr float rotor_inertia {2670}; // in gcm2
   };
 
   /**\class X8ProHV3
@@ -170,10 +170,10 @@ namespace myactuator_rmd {
   */
   class X8ProHV3 {
     public:
-      static constexpr double reducer_ratio {6.2/1};
-      static constexpr double speed_constant {30}; // in rpm/V
-      static constexpr double torque_constant {0.29}; // in Nm/A
-      static constexpr double rotor_inertia {3400}; // in gcm2
+      static constexpr float reducer_ratio {6.2/1};
+      static constexpr float speed_constant {30}; // in rpm/V
+      static constexpr float torque_constant {0.29}; // in Nm/A
+      static constexpr float rotor_inertia {3400}; // in gcm2
   };
 
 }
